Refactors testing files into Test_Game and Test_Pieces for ease of reading.
